make customize/edit routine more straightforward
